Unit tests to test signup create/update endpoint
