+++ registrations/tests/test_registration_admin_side.py
@@ -38,66 +38,6 @@ def get_event_url(detail_pk):
 
 
 @pytest.mark.django_db
-def test_successful_sign_up(api_client, languages, registration):
-    """User reserves seats and then signs up."""
-    sign_up_data = {
-        "name": "Michael Jackson",
-        "date_of_birth": "2011-04-07",
-        "email": "test@test.com",
-        "phone_number": "0441111111",
-        "notifications": "sms",
-        "service_language": "fi",
-        "native_language": "fi",
-        "street_address": "my street",
-        "zipcode": "myzip1",
-    }
-
-    response = sign_up(api_client, registration.id, sign_up_data)
-
-    assert response.status_code == status.HTTP_201_CREATED
-    assert SignUp.objects.count() == 1
-
-    signup = SignUp.objects.first()
-    assert signup.attendee_status == SignUp.AttendeeStatus.ATTENDING
-    assert signup.name == sign_up_data["name"]
-    assert signup.date_of_birth == date(2011, 4, 7)
-    assert signup.email == sign_up_data["email"]
-    assert signup.phone_number == sign_up_data["phone_number"]
-    assert signup.notifications == SignUp.NotificationType.SMS
-    assert signup.native_language.pk == "fi"
-    assert signup.service_language.pk == "fi"
-    assert signup.street_address == sign_up_data["street_address"]
-    assert signup.zipcode == sign_up_data["zipcode"]
-
-
-@pytest.mark.django_db
-def test_cannot_sign_up_twice_with_same_phone_or_email(api_client, registration):
-    sign_up_data = {
-        "name": "Michael Jackson",
-        "email": "test@test.com",
-        "phone_number": "0441111111",
-        "date_of_birth": "2011-04-07",
-    }
-
-    response = sign_up(api_client, registration.id, sign_up_data)
-    assert response.status_code == status.HTTP_201_CREATED
-
-    # cannot signup with the same email twice
-    sign_up_data_same_email = deepcopy(sign_up_data)
-    sign_up_data_same_email["phone_number"] = "0442222222"
-    response = sign_up(api_client, registration.id, sign_up_data_same_email)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.data["non_field_errors"][0].code == "unique"
-
-    # cannot signup with the same phone twice
-    sign_up_data_same_phone = deepcopy(sign_up_data)
-    sign_up_data_same_phone["email"] = "another@email.com"
-    response = sign_up(api_client, registration.id, sign_up_data_same_phone)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.data["non_field_errors"][0].code == "unique"
-
-
-@pytest.mark.django_db
 def test_current_attendee_and_waitlist_count(api_client, user, event):
     registration_url = reverse("registration-list")
     registration_data = {
@@ -144,135 +84,6 @@ def test_current_attendee_and_waitlist_count(api_client, user, event):
     response = api_client.get(registration_detail_url, format="json")
     assert response.data["current_attendee_count"] == 1
     assert response.data["current_waiting_list_count"] == 1
-
-
-@pytest.mark.parametrize("min_age", [None, 0, 10])
-@pytest.mark.parametrize("max_age", [None, 0, 100])
-@pytest.mark.parametrize("date_of_birth", [None, "1980-12-30"])
-@freeze_time("2023-03-14 03:30:00+02:00")
-@pytest.mark.django_db
-def test_signup_age_is_mandatory_if_audience_min_or_max_age_specified(
-    api_client, user, event, date_of_birth, min_age, max_age
-):
-    falsy_values = ("", None)
-
-    if (
-        min_age not in falsy_values or max_age not in falsy_values
-    ) and not date_of_birth:
-        expected_status = status.HTTP_400_BAD_REQUEST
-        expected_error = "This field must be specified."
-    else:
-        expected_status = status.HTTP_201_CREATED
-        expected_error = None
-
-    registration_url = reverse("registration-list")
-    sign_up_data = {
-        "name": "Michael Jackson",
-        "email": "test@test.com",
-        "phone_number": "0441111111",
-        "notifications": "sms",
-    }
-    if date_of_birth:
-        sign_up_data["date_of_birth"] = date_of_birth
-
-    # Create registration
-    api_client.force_authenticate(user)
-    registration_data = {
-        "event": {"@id": get_event_url(event.pk)},
-        "maximum_attendee_capacity": 1,
-    }
-    if min_age not in falsy_values:
-        registration_data["audience_min_age"] = min_age
-    if max_age not in falsy_values:
-        registration_data["audience_max_age"] = max_age
-    response = api_client.post(registration_url, registration_data, format="json")
-    registration_id = response.data["id"]
-
-    response = sign_up(api_client, registration_id, sign_up_data)
-
-    assert response.status_code == expected_status
-    if expected_error:
-        assert str(response.data["date_of_birth"][0]) == expected_error
-
-
-@pytest.mark.parametrize(
-    "date_of_birth,expected_status,expected_error",
-    [
-        ("2011-04-07", status.HTTP_400_BAD_REQUEST, "The participant is too young."),
-        ("1879-03-14", status.HTTP_400_BAD_REQUEST, "The participant is too old."),
-        ("2000-02-29", status.HTTP_201_CREATED, None),
-    ],
-)
-@freeze_time("2023-03-14 03:30:00+02:00")
-@pytest.mark.django_db
-def test_signup_age_has_to_match_the_audience_min_max_age(
-    api_client, user, event, date_of_birth, expected_status, expected_error
-):
-    registration_url = reverse("registration-list")
-    registration_data = {
-        "event": {"@id": get_event_url(event.pk)},
-        "audience_max_age": 40,
-        "audience_min_age": 20,
-        "maximum_attendee_capacity": 1,
-    }
-    sign_up_data = {
-        "name": "Michael Jackson",
-        "email": "test@test.com",
-        "phone_number": "0441111111",
-        "notifications": "sms",
-        "date_of_birth": date_of_birth,
-    }
-
-    api_client.force_authenticate(user)
-    response = api_client.post(registration_url, registration_data, format="json")
-    registration_id = response.data["id"]
-
-    response = sign_up(api_client, registration_id, sign_up_data)
-
-    assert response.status_code == expected_status
-    if expected_error:
-        assert str(response.data["date_of_birth"][0]) == expected_error
-
-
-@pytest.mark.parametrize(
-    "mandatory_field_id",
-    [
-        MandatoryFields.CITY,
-        MandatoryFields.NAME,
-        MandatoryFields.PHONE_NUMBER,
-        MandatoryFields.STREET_ADDRESS,
-        MandatoryFields.ZIPCODE,
-    ],
-)
-@pytest.mark.django_db
-def test_signup_mandatory_fields_has_to_be_filled(
-    api_client, event, mandatory_field_id, user
-):
-    api_client.force_authenticate(user)
-
-    registration_url = reverse("registration-list")
-    registration_data = {
-        "event": {"@id": get_event_url(event.pk)},
-        "mandatory_fields": [mandatory_field_id],
-    }
-    response = api_client.post(registration_url, registration_data, format="json")
-    assert response.status_code == status.HTTP_201_CREATED
-    registration_id = response.data["id"]
-
-    sign_up_data = {
-        "name": "Michael Jackson",
-        "email": "test@test.com",
-        "phone_number": "0441111111",
-        "street_address": "Street address",
-        "city": "Helsinki",
-        "zipcode": "00100",
-        "notifications": "sms",
-    }
-    sign_up_data[mandatory_field_id] = ""
-
-    response = sign_up(api_client, registration_id, sign_up_data)
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert str(response.data[mandatory_field_id][0]) == "This field must be specified."
 
 
 @pytest.mark.skip(
